+++ tests/testthat/test-ifcb_annotate_samples.R
@@ -1,6 +1,5 @@
 test_that("ifcb_annotate_samples correctly updates the .mat classlist files", {
   # Skip if scipy is not available
-  skip_if_no_scipy()
 
   # Define the path to the test data zip file
   zip_path <- test_path("test_data/test_data.zip") # Path to the test data zip file containing .mat files and config
@@ -47,7 +46,6 @@ test_that("ifcb_annotate_samples correctly updates the .mat classlist files", {
 
 test_that("ifcb_annotate_samples correctly updates the .mat classlist files if classes are unmatches", {
   # Skip if scipy is not available
-  skip_if_no_scipy()
 
   # Define the path to the test data zip file
   zip_path <- test_path("test_data/test_data.zip") # Path to the test data zip file containing .mat files and config
@@ -105,7 +103,6 @@ test_that("ifcb_annotate_samples correctly updates the .mat classlist files if c
 
 test_that("ifcb_annotate_samples handles missing directories and files gracefully", {
   # Skip if scipy is not available
-  skip_if_no_scipy()
 
   # Define the path to the test data zip file
   zip_path <- test_path("test_data/test_data.zip") # Path to the test data zip file containing .mat files and config