+++ tests/testthat/test-ifcb_prepare_whoi_plankton.R
@@ -1,6 +1,5 @@
 test_that("ifcb_prepare_whoi_plankton works", {
   # Skip if Python is not available
-  skip_if_no_scipy()
 
   # Skip if offline
   skip_if_offline()
@@ -121,7 +120,6 @@ test_that("ifcb_prepare_whoi_plankton works", {
 
 test_that("ifcb_prepare_whoi_plankton throws errors", {
   # Skip if Python is not available
-  skip_if_no_scipy()
 
   # Check error
   expect_error(ifcb_prepare_whoi_plankton(2,